requirements.txt
Flask==3.1.0
Flask_Cors==4.0.0
gspread==6.1.4
langchain==0.3.13
langchain_chroma==0.1.4
langchain_core==0.3.28
langchain_huggingface==0.1.2
langchain_ollama==0.2.2
pandas==2.2.3
protobuf==5.29.2
python-dotenv==1.0.1
flask-apscheduler
